requirements.txt
Flask
Flask-WTF
requests
python-dotenv==1.0.0